feat: added testing cases for current modules
